refactor: 견적 요청 페이지 import 경로 및 세부 디자인 수정
